refactoring: extract hierarchy builder
