+++ src/js/processing/index.js
@@ -1,0 +1,2 @@
+export {buildHierarchy} from './hierarchy';
+export {processTimeMapData} from './timemap';
